refactor (ui): move Heading to components/ui with barrel exports and remove Typography and Buttons route
