--- frontend/src/components/typography/Typography.svelte
@@ -1,1 +1,0 @@
-<p><slot /></p>
