--- frontend/src/components/typography/index.ts
@@ -1,2 +1,0 @@
-export { default as Heading } from './Heading.svelte';
-export { default as Typography } from './Typography.svelte';
