+++ frontend/src/components/ui/index.ts
@@ -1,2 +1,2 @@
-// ui/index.ts
 export { default as Button } from './Button.svelte';
+export { default as Heading } from './Heading.svelte';
